requirements.txt
asgiref==3.9.2
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
inquirerpy==0.3.4
packaging==25.0
pfzy==0.3.4
prompt_toolkit==3.0.52
PyJWT==2.10.1
PyYAML==6.0.3
requests==2.32.5
setuptools==78.1.1
simplejson==3.20.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
wheel==0.45.1
